Add css styling
There is a new folder added names css in the static folder. The
folder will consist of all the css files that are required in the
project. There is also a file added styles.css that has some styling
added to it. The styling changes the design and layout of home.html
and layout.html at the moment.